+++ eva/src/components/dashBoardInside/dashMap.vue
@@ -421,10 +421,13 @@ export default {
             // создаем элемент карты
             this.map.createMap(this.maptheme);
             // рисуем объекты на карте
-            this.map.drawObjects(
+            this.map.drawObjects({
               dataRest,
-              this.pipelineDataDictionary,
-            );
+              pipelineDataDictionary: this.pipelineDataDictionary,
+              callback: (id) => {
+                this.setClick(id);
+              },
+            });
             if (this.map) {
               if (this.options.initialPoint) {
                 this.map.setView(
@@ -556,6 +559,51 @@ export default {
           this.map.resize();
         }
       });
+    },
+
+    setClick(tokenValue) {
+      const events = this.getEvents({
+        event: 'onclick',
+        partelement: 'empty',
+      });
+      if (events.length !== 0) {
+        events.forEach((item) => {
+          if (item.action === 'go') {
+            item.value[0] = tokenValue;
+            this.$store.dispatch('letEventGo', {
+              event: item,
+              id: this.element,
+              idDash: this.idDash,
+              route: this.$router,
+              store: this.$store,
+            });
+          }
+        });
+      }
+    },
+    getEvents({ event, partelement }) {
+      let result = [];
+      if (!this.$store.state[this.idDash].events) {
+        this.$store.commit('setState', [{
+          object: this.$store.state[this.idDash],
+          prop: 'events',
+          value: [],
+        }]);
+        return [];
+      }
+      if (partelement) {
+        result = this.$store.state[this.idDash].events.filter((item) => (
+          item.event === event
+            && item.element === this.element
+            && item.partelement === partelement
+        ));
+      } else {
+        result = this.$store.state[this.idDash].events.filter(
+          (item) => item.event === event
+                && item.target === this.element,
+        );
+      }
+      return result;
     },
   },
 };